requirements.txt
flask
Flask-Session
Flask-Caching
PyJWT
dash
dash-bootstrap-components
dash-extensions
python-dotenv
pandas
